Use last ranking data on tags pages
